big update

+++ sam2/ops/modules/ms_deform_attn.py
@@ -111,14 +111,14 @@ class MSDeformAttn(nn.Module):
                 'Last dim of reference_points must be 2 or 4, but get {} instead.'.format(reference_points.shape[-1]))
 
         # for amp
-        if value.dtype == torch.float16:
+        if value.dtype == torch.bfloat16:
             # for mixed precision
             output = MSDeformAttnFunction.apply(
             value.to(torch.float32), input_spatial_shapes, input_level_start_index, sampling_locations.to(torch.float32), attention_weights, self.im2col_step)
             output = output.to(torch.float16)
             output = self.output_proj(output)
             return output
-
+        
         output = MSDeformAttnFunction.apply(
             value, input_spatial_shapes, input_level_start_index, sampling_locations, attention_weights, self.im2col_step)
